Move entities to data
