+++ android/app/src/androidTest/java/io/github/kn65op/domag/dbtests/CategoryLimitDatabaseTest.kt
@@ -2,7 +2,7 @@ package io.github.kn65op.domag.dbtests
 
 import io.github.kn65op.android.lib.type.FixedPointNumber
 import io.github.kn65op.domag.data.database.daos.CategoryLimitDao
-import io.github.kn65op.domag.data.database.entities.CategoryLimit
+import io.github.kn65op.domag.data.entities.CategoryLimit
 import io.github.kn65op.domag.dbtests.common.DatabaseTest
 import io.github.kn65op.domag.dbtests.common.assertNoData
 import io.github.kn65op.domag.dbtests.common.getFromLiveData
