+++ android/app/src/main/java/io/github/kn65op/domag/data/database/daos/ConsumeDao.kt
@@ -2,7 +2,7 @@ package io.github.kn65op.domag.data.database.daos
 
 import androidx.lifecycle.LiveData
 import androidx.room.*
-import io.github.kn65op.domag.data.database.entities.Consume
+import io.github.kn65op.domag.data.entities.Consume
 
 @Dao
 interface ConsumeDao {
