+++ android/app/src/main/java/io/github/kn65op/domag/data/database/relations/DepotWithContents.kt
@@ -2,8 +2,8 @@ package io.github.kn65op.domag.data.database.relations
 
 import androidx.room.Embedded
 import androidx.room.Relation
-import io.github.kn65op.domag.data.database.entities.Item
-import io.github.kn65op.domag.data.database.entities.Depot
+import io.github.kn65op.domag.data.entities.Item
+import io.github.kn65op.domag.data.entities.Depot
 
 data class DepotWithContents(
     @Embedded
