+++ android/app/src/main/java/io/github/kn65op/domag/data/database/utils/ItemsUtils.kt
@@ -1,7 +1,7 @@
 package io.github.kn65op.domag.data.database.utils
 
 import io.github.kn65op.android.lib.type.FixedPointNumber
-import io.github.kn65op.domag.data.database.entities.Item
+import io.github.kn65op.domag.data.entities.Item
 
 fun List<Item>.amount() = when (val wholeAmount =
     map { it.amount }.reduceOrNull { acc, itemAmount -> acc + itemAmount }) {
