+++ android/app/src/main/java/io/github/kn65op/domag/ui/depot/EditDepotFragment.kt
@@ -14,7 +14,7 @@ import com.toptoche.searchablespinnerlibrary.SearchableSpinner
 import io.github.kn65op.domag.R
 import io.github.kn65op.domag.data.database.database.AppDatabase
 import io.github.kn65op.domag.data.database.database.DatabaseFactoryImpl
-import io.github.kn65op.domag.data.database.entities.Depot
+import io.github.kn65op.domag.data.entities.Depot
 import io.github.kn65op.domag.data.database.operations.deleteDepot
 import io.github.kn65op.domag.data.database.relations.DepotWithContents
 import io.github.kn65op.domag.databinding.FragmentEditDepotBinding
