+++ android/app/src/main/java/io/github/kn65op/domag/ui/items/ItemsFragment.kt
@@ -11,7 +11,7 @@ import androidx.recyclerview.widget.RecyclerView
 import io.github.kn65op.domag.R
 import io.github.kn65op.domag.data.database.database.AppDatabase
 import io.github.kn65op.domag.data.database.database.DatabaseFactoryImpl
-import io.github.kn65op.domag.data.database.entities.Depot
+import io.github.kn65op.domag.data.entities.Depot
 import io.github.kn65op.domag.data.database.relations.DepotWithContents
 import io.github.kn65op.domag.ui.common.FragmentWithActionBar
 import io.github.kn65op.domag.ui.common.prepareFabs
